python/requirements.txt
# Deep Learning Framework with GPU support
tensorflow>=2.15.0
keras>=3.0.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Model utilities
scikit-learn>=1.3.0

# API Framework
flask>=3.0.0
flask-cors>=4.0.0

# Progress tracking
tqdm>=4.66.0
